Fix exception handling
