+++ src/vm/mod.rs
@@ -71,11 +71,26 @@ pub fn run(arguments: &Arguments, java_home: &Path) -> Result<()> {
             Ok(())
         }
         Err(e) if e.is_uncaught_exception() => {
+            if let Some(throwable_ref) = e.throwable_ref() {
+                invoke_uncaught_exception_handler(throwable_ref)?;
+            }
             invoke_shutdown_hooks()?;
             Err(e)
         }
         Err(e) => Err(e),
     }
+}
+
+fn invoke_uncaught_exception_handler(throwable_ref: i32) -> Result<()> {
+    let system_thread_group_ref = with_method_area(|area| area.system_thread_group_id())?;
+    let system_thread_ref = with_method_area(|area| area.system_thread_id())?;
+    Executor::invoke_non_static_method(
+        "java/lang/ThreadGroup",
+        "uncaughtException:(Ljava/lang/Thread;Ljava/lang/Throwable;)V",
+        system_thread_group_ref,
+        &[system_thread_ref.into(), throwable_ref.into()],
+    )?;
+    Ok(())
 }
 
 fn invoke_shutdown_hooks() -> Result<()> {
